creates DM anon reports
