+++ src/bootstrap/plugin.loader.ts
@@ -23,6 +23,7 @@ import { ScoresPlugin } from '../app/plugins/scores.plugin';
 import { MarketPlacePlugin } from '../app/plugins/marketplace.plugin';
 import { ManageRolesPlugin } from '../app/plugins/manageroles.plugin';
 import { StatusPlugin } from '../app/plugins/status.plugin';
+import { DmReportPlugin } from '../app/plugins/anonreport.plugin';
 
 const PluginStore: { [pluginName: string]: any } = {
   dog: DogPlugin,
@@ -50,6 +51,7 @@ const PluginStore: { [pluginName: string]: any } = {
   marketplace: MarketPlacePlugin,
   manageroles: ManageRolesPlugin,
   status: StatusPlugin,
+  anonreport: DmReportPlugin,
 };
 
 export class PluginLoader {
